requirements.txt
asgiref==3.8.1
Django==4.2.20
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.0
markdown>=3.8
pymysql
drf-yasg>=1.21.7
